separate alonzo tests
